⚒️ chore: add ExchangeRate repo
